AI/requirements.txt
# AI/requirements.txt

# ============================================================================
# Core Framework
# ============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# ============================================================================
# ASR - faster-whisper (Dockerfile에서 별도 설치)
# ============================================================================
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.10.0
webrtcvad>=2.0.10

# ============================================================================
# NLU & LLM (외부 API)
# ============================================================================
openai>=1.12.0

# ============================================================================
# TTS - Google Cloud TTS
# ============================================================================
google-cloud-texttospeech>=2.16.0
httpx>=0.26.0
aiohttp>=3.9.0

# ============================================================================
# OCR - PaddleOCR (Dockerfile에서 별도 설치)
# ============================================================================
Pillow>=10.2.0

# PaddleOCR (한국어 OCR)
# ⚠️ PaddlePaddle은 CUDA 버전에 따라 별도 설치 필요:
#   CUDA 11.8: python -m pip install paddlepaddle-gpu==3.0.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu118/
#   CUDA 12.6: python -m pip install paddlepaddle-gpu==3.0.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu126/
#   CPU Only:  python -m pip install paddlepaddle==3.0.0 -i https://www.paddlepaddle.org.cn/packages/stable/cpu/
paddleocr>=2.7.0

# 사용 모델: korean_PP-OCRv5_mobile_rec (한국어 텍스트 인식)

# ============================================================================
# WebSocket & Async
# ============================================================================
websockets>=12.0

# ============================================================================
# Utilities
# ============================================================================
colorlog>=6.8.0
python-dateutil>=2.8.2
